feat: update kanban lane

+++ src/main/java/com/projectboated/backend/domain/kanban/kanbanlane/service/KanbanLaneService.java
@@ -7,16 +7,17 @@ import com.projectboated.backend.domain.common.exception.ErrorCode;
 import com.projectboated.backend.domain.kanban.kanbanlane.entity.KanbanLane;
 import com.projectboated.backend.domain.kanban.kanbanlane.repository.KanbanLaneRepository;
 import com.projectboated.backend.domain.kanban.kanbanlane.service.dto.ChangeTaskOrderRequest;
-import com.projectboated.backend.domain.kanban.kanbanlane.service.exception.KanbanLaneAlreadyExists5;
-import com.projectboated.backend.domain.kanban.kanbanlane.service.exception.KanbanLaneSaveAccessDeniedException;
-import com.projectboated.backend.domain.kanban.kanbanlane.service.exception.TaskChangeIndexDeniedException;
+import com.projectboated.backend.domain.kanban.kanbanlane.service.dto.KanbanLaneUpdateRequest;
+import com.projectboated.backend.domain.kanban.kanbanlane.service.exception.*;
 import com.projectboated.backend.domain.project.entity.Project;
 import com.projectboated.backend.domain.project.repository.ProjectRepository;
-import com.projectboated.backend.domain.project.service.AccountProjectService;
+import com.projectboated.backend.domain.project.service.ProjectService;
 import com.projectboated.backend.domain.project.service.exception.ProjectNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -25,7 +26,7 @@ public class KanbanLaneService {
     private final AccountRepository accountRepository;
     private final ProjectRepository projectRepository;
     private final KanbanLaneRepository kanbanLaneRepository;
-    private final AccountProjectService accountProjectService;
+    private final ProjectService projectService;
 
     @Transactional
     public void createNewLine(Long accountId, Long projectId, String name) {
@@ -52,7 +53,7 @@ public class KanbanLaneService {
     }
 
     @Transactional
-    public void deleteCustomLane(Long accountId, Long projectId, Long kanbanLaneId) {
+    public void deleteKanbanLane(Long accountId, Long projectId, Long kanbanLaneId) {
         Account account = accountRepository.findById(accountId)
                 .orElseThrow(() -> new AccountNotFoundException(ErrorCode.ACCOUNT_NOT_FOUND));
 
@@ -61,6 +62,10 @@ public class KanbanLaneService {
 
         if(!project.isCaptain(account)) {
             throw new KanbanLaneSaveAccessDeniedException(ErrorCode.PROJECT_ONLY_CAPTAIN);
+        }
+
+        if (!isInProject(kanbanLaneId, project)) {
+            throw new KanbanLaneNotInProjectException(ErrorCode.KANBAN_LANE_NOT_IN_PROJECT);
         }
 
         kanbanLaneRepository.deleteById(kanbanLaneId);
@@ -74,11 +79,43 @@ public class KanbanLaneService {
         Project project = projectRepository.findById(request.projectId())
                 .orElseThrow(() -> new ProjectNotFoundException(ErrorCode.PROJECT_NOT_FOUND));
 
-        if (!project.isCaptain(account) &&
-                !accountProjectService.isCrew(project, account)) {
-            throw new TaskChangeIndexDeniedException(ErrorCode.PROJECT_ONLY_CAPTAIN_OR_CREW);
+        if (!projectService.isCaptainOrCrew(project, account)) {
+            throw new TaskChangeOrderAccessDeniedException(ErrorCode.PROJECT_ONLY_CAPTAIN_OR_CREW);
         }
 
         project.getKanban().changeTaskOrder(request);
+    }
+
+    @Transactional
+    public void updateKanbanLane(Long accountId, KanbanLaneUpdateRequest request) {
+        Account account = accountRepository.findById(accountId)
+                .orElseThrow(() -> new AccountNotFoundException(ErrorCode.ACCOUNT_NOT_FOUND));
+
+        Project project = projectRepository.findById(request.getProjectId())
+                .orElseThrow(() -> new ProjectNotFoundException(ErrorCode.PROJECT_NOT_FOUND));
+
+        if (!project.isCaptain(account)) {
+            throw new KanbanLaneUpdateAccessDeniedException(ErrorCode.PROJECT_ONLY_CAPTAIN);
+        }
+
+        project.getKanban().updateLane(request.getKanbanLaneIndex(), request);
+    }
+
+    private boolean isInProject(KanbanLane kanbanLane, Project project) {
+        for (KanbanLane lane : project.getKanban().getLanes()) {
+            if (lane.getId() == kanbanLane.getId()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean isInProject(Long kanbanLaneId, Project project) {
+        for (KanbanLane lane : project.getKanban().getLanes()) {
+            if (Objects.equals(lane.getId(), kanbanLaneId)) {
+                return true;
+            }
+        }
+        return false;
     }
 }
